Oban like migration system
This commit introduces an Oban like migration system for the ErrorTracker.
Users can generate their own migrations that call the ErrorTracker
migrations and possibly set an alternative database prefix in case that
they want to store the errors in a separate schema than the rest of the
application.

This commit also updates the dev script to showcase how to use the new
migration system on an alternative prefix.