Do not fingerprint error reason (#19)
Including the error reason among the fingerprinted parameters caused wrong error grouping in certain scenarios such as accessing not-existing map fields.

This pull request stores only the original reason in the error itself and then records each occurrence reason. This ensures that errors are grouped together even though the reason has slight changes.

I am not sure if it is worth to keep the reason in the error itself, but we can always remove it later so this doesn't have to be decided now.