Add SQLite3 support ans `meta` table (#31)
This change adds support for SQLite3 storage of errors and includes a new `error_tracker_meta` table to store metadata needed for ErrorTracker to function properly - like managing migration versions.

Aside from adding support to SQLite3, which was as easy as expected, we have also added a way to track migration versions that should work in any SQL database (although only PostgreSQL and SQLite3 are currently supported). That KV store table will also work in the future to store runtime configuration, for example.

In this change we are also migrating PostgreSQL systems to use this new table, which means migrating from using table comments to store migration version to use this new table. 

As for SQLite3 migrations, migrations start at `V02` to maintain the same version IDs on all platforms.