Unit tests (#34)
Adds unit tests for:
- `ErrorTracker`.
  - `report/3`
    - with a runtime exception
    - with a bad arithmetic error
    - with an unknown function error
    - with a throw
    - with an exit
  - `resolve/1`
  - `unresolve/1`
- `Telemetry`
  - verifies that events are emitted

Tests can run either on PostgreSQL or SQLite3. This is determined by the `DB` env var so:
- `DB=postgres mix test` runs the tests on PostgreSQL
- `DB=sqlite mix test` runs the tests on SQLite3

The CI has been also updated to run tests on both databases.

---

I've also unified the configuration under the `config/` folder. And the dev/test migrations under `priv/repo/`.

Previously we had some configuration in the root folder and other in `config/`. Regarding the migrations the `priv/repo/` folder is not included when publishing the package but is used in dev/test.