Fix error search in SQLite (#51)
SQLite does not support the ILIKE operator. We need to use the LIKE
operator instead, which performs a case-insensitive match for ASCII
characters.

Since we are looking at stacktrace lines and module names I think that
it is fair to expect ASCII characters 99%+ of the time anyway.

This pull request fixes the problem by detecting the current database
adapter and using LIKE or ILIKE accordingly. I've also extracted the
search function out of the dashboard and unit-tested it to ensure that
it works and doesn't break by mistake.

Closes #50