Use explicit column name for FK references
The ErrorTracker tables use `id` as the primary key name. This was
causing trouble when the client repository was configured to use a
different column name for references since it tried to create a foreign
key pointing to a non-existing column.

This commit explicitly states the referenced column name to avoid this
problems.