Do not store error if context/occurrence is not valid (#66)
This commit fixes an issue when providing invalid content inside the
error context. It also updates the documentation to explicitly state
that the context data must be JSON-encodeable.

Previously when providing incorrect context data the Error would be
stored but not the Occurrence. This produced Errors that show up on the
dashboard but cannot be opened because they have no Occurrences.

This commit updates the process to use a DB transaction so it ensures
that both the Error and Occurrence are stored properly, or not at all.

Closes #65