Fix Oban reporting and all state field.
There was an error when an Oban worker returned an `{:error, reason}`
tuple because Oban does not return an stacktrace (makes sense, as
it is not an exception by itself).

Our code was not prepared for that use case. I modified it to handle
it and circunvent the needed for source file and source function
information.

As an extra feature I have also added the reporting of the `state`
of the job (which is mostly ever `failure` for us).