Store Ash bread crumbs in occurrence context (#106)
Some frameworks such as Ash include breadcrumbs in their errors. This is
provided by the fantastic Splode library
([details](https://bsky.app/profile/zachdaniel.dev/post/3l7qsyil4pg2c)).

This commit updates the ErrorTracker to detect such bread crumbs and
include them in the occurrence context. In the future we may want to
revisit this and store them in their own field separate from the
context.

It is important to note that you don't have to use Ash or Splode to
leverage this integration. Breadcrumbs will be automatically extracted
from any exception that includes a `bread_crumbs` field.

I've discussed this with @zachdaniel and he kindly integrated the
breadcrumbs on Ash main for create and change actions.