Detect changes to static asset files
A common problem during development is that we make changes to the CSS
or JS files and restart the dev server only to find that those changes
haven't been picked up.

This happens because we are reading those files during compilation but,
after the layout is compiled, newer changes are not picked up.

This commit fixes this by using the `@external_resource` attribute,
which tells the Elixir compiler that the module depends on external
files and must be recompiled when those change.

This fixes our problem and ensures that our dev server always uses the
updated assets when we make changes.