Improve documentation of ErrorTracker.Error kind field (#154)
We thought about it and came to the conclusion that it would be good to
change the type declaration for the module. Open for discussion, because
that could potentially be a breaking change of the library if users
previously had implementations of `ErrorTracker.Ignorer` that matched on
non-string types (although, arguably those implementations would never
have worked).

This addresses #153.

---------

Co-authored-by: Manmohan Krishna <manmohan.krishna@maersk.com>